Added set file function
getData will now auto load a file and add it to the pipeline after
flusing it
