+++ Software/waveview/scope_link/include/EVTester.hpp
@@ -13,8 +13,6 @@ void testTriggerThroughput();
 
 void testCsv(char * filename);
 
-bool loadFromFile ( char* filename, boost::lockfree::queue<buffer*, boost::lockfree::fixed_sized<false>> *outputQ);
-
 void runSocketTest();
 
 #endif /* EVTester_hpp */
