+++ Software/waveview/scope_link/src/EVTester.cpp
@@ -12,63 +12,6 @@ uint32_t testSize = 1000;
 
 Bridge* bridgeThread_1;
 
-bool loadFromFile ( char* filename, boost::lockfree::queue<buffer*, boost::lockfree::fixed_sized<false>> *outputQ)
-{
-    std::ifstream stream;
-    stream.open(filename);
-
-    char delim = '\n';
-    std::string tmp;
-
-    INFO << "Loading from file " << filename;
-
-    buffer* tempBuffer;
-    tempBuffer = bufferAllocator.allocate(1);
-    bufferAllocator.construct(tempBuffer);
-    uint32_t tmpBufPos = 0;
-
-    if (!stream.is_open()) {
-        ERROR << "Stream is closed. Expects relative path. Run from waveview folder";
-        return false;
-    }
-
-    while (std::getline(stream, tmp, delim)) {
-//        INFO << "Parsing line into buffer";
-        // Parse the line into a buffer
-        typedef boost::tokenizer<boost::char_separator<char>> tokenizer;
-        boost::char_separator<char> sep{","};
-        tokenizer tok{tmp, sep};
-        for (const auto &t : tok) {
-
-            if (std::stoi(t) > 255) {
-                ERROR << "Number greater than 255";
-            } else if (std::stoi(t) > INT8_MAX) {
-                ERROR << "Number greater than 127 is converted to negative";
-            } else if ((int8_t)std::stoi(t) < -128) {
-                ERROR << "Number less than -128";
-            }
-
-            tempBuffer->data[tmpBufPos] = (int8_t)std::stoi(t);
-
-            tmpBufPos++;
-            if (tmpBufPos == BUFFER_SIZE) {
-                DEBUG << "Adding buffer to queue from file";
-                // Buffer is now full push it
-                outputQ->push(tempBuffer);
-
-                // Create a new one to fill
-                tempBuffer = bufferAllocator.allocate(1);
-                bufferAllocator.construct(tempBuffer);
-                tmpBufPos = 0;
-            }
-        }
-    }
-    // Deleted any partially filled buffer
-    bufferAllocator.deallocate(tempBuffer, 1);
-
-    return true;
-}
-
 void loadFromRand (boost::lockfree::queue<buffer*, boost::lockfree::fixed_sized<false>> *outputQ)
 {
     // Create dummy buffer
