+++ Software/waveview/scope_link/src/postProcessor.cpp
@@ -136,7 +136,7 @@ void postProcessor::coreLoop()
             // flush the input queue so it doesn't overflow
             size_t count = 0;
             count = (*inputQueue).consume_all(free);
-//            DEBUG << "Flushed postProcessor inputQueue: " << count;
+            TRACE << "Flushed postProcessor inputQueue: " << count;
         }
         // Queue empty, Sleep for a bit
         std::this_thread::sleep_for(std::chrono::microseconds(100));
