+++ Software/waveview/scope_link/src/processor.cpp
@@ -191,7 +191,7 @@ void Processor::coreLoop()
             // flush the input queue so it doesn't overflow
             size_t count = 0;
             count = (*inputQueue).consume_all(bufferFunctor);
-//            DEBUG << "Flushed processor inputQueue: " << count;
+            TRACE << "Flushed processor inputQueue: " << count;
         }
         // Queue empty, Sleep for a bit
         std::this_thread::sleep_for(std::chrono::microseconds(100));
